requirements.txt
apache-airflow-providers-postgres[amazon]
python-decouple==3.6
python-dotenv==0.21.0